requirements.txt
databricks-sdk
streamlit
duckdb
gradio
gradio[mcp]